chore: 基线——本地适配改动（item_code 255 / row_count BIGINT / 提资复核分批流转）
本地改动基线（ui-refactor 分支起点），来源见 04-项目/deshu5/本地适配与修复记录.md §一。此后前端重构按工作项小步提交；仅本地，永不 push。

+++ core/schema_kb.py
@@ -88,7 +88,7 @@ class SchemaKnowledgeBase:
                     id INT PRIMARY KEY AUTO_INCREMENT,
                     table_name VARCHAR(128) UNIQUE,
                     table_comment VARCHAR(255),
-                    row_count INT,
+                    row_count BIGINT,
                     column_count INT,
                     doc_text TEXT,
                     updated_at DATETIME DEFAULT CURRENT_TIMESTAMP
